Implement better Enderman mobGriefing game rule
Add new "mobGriefingEnderman" game rule, this will be a boolean value
and will override the default "mobGriefing" game rule when the entity is
an Enderman.
This rule should be added on world start.

Create core mod framework to handle scenarios which cannot be covered by
Forge/FML events.
The core mod will register a class transformer to allow mobGriefing
rules to be modified for a particular class.

Add EntityEnderman class transform to the class transformer, replacing
"mobGriefing" with "mobGriefingEnderman". This will control the ability
for Enderman to pick up and place blocks.

Close #3