Fix "mobGriefingCreeper" not working at runtime
The "mobGriefingCreeper" game rule does not correctly work at runtime,
this is due to the handler incorrectly using a MCP field name to
identify the "isSmoking" field.

Change the event handler onDetonatEvent to pass both SRG and MCP naming
to ObfuscationReflectionHelper to allow access to the correct field at
both runtime and within the development environment.

Close #1