Implement better Ghast mobGriefing game rule
Add new "mobGriefingGhast" game rule, this will be a boolean value and
will override the default "mobGriefing" game rule when the entity is a
Ghast.

Modify the event handler's onDetonateEvent method to support explosions
from Ghast fireballs.
The Explosion object's exploder field does not get populated for Ghast
fireballs, instead the a EntityLargeFireball object will be included in
the affected objects field.

The affected objects will be iterated to try and identify a fireball
object belonging to a Ghast.
If found and the Ghast's fireball is the cause of the explosion event
then the new "mobGriefingGhast" game rule will be used to determine
whether the explosion will do block damage.

Close #4