Implement better Zombie mobGriefing game rule
Add new "mobGriefingZombie" game rule, this will be a boolean value and
will override the default "mobGriefing" game rule when the entity is a
Zombie.

Extend the EntityAIBreakDoor AI task, the extended task will override
the "shouldExecute" method and use the new mobGriefing rule to determine
whether the task should be executed or not.

Add a new method to BetterMobGriefingGameRuleEventHandler to handle
EntityJoinWorldEvent. When this event is received the field
"field_146075_bs" will be updated to contain an instance of the new
EntityAIBreakDoor extension.
The entity's task list will also be updated if the original
EntityAIBreakDoor task has already been populated in task list.

Update .gitignore to ignore the "logs" folder, which is created when
running Unit Tests.

Close #8