Implement better Sheep mobGriefing game rule
Add new "mobGriefingSheep" game rule, this will be a boolean value and
will override the default "mobGriefing" game rule when the entity is
a EntitySheep.

Add to the EntityJoinWorldEvent handler, the entity's task list will be
checked for any instances of EntityAIEatGrass. If any are found they
will be replaced with a new AI task which will use the better
mobGriefing game rules.

Close #9