Implement configuration for default values
Implement a mod configuration to allow specification of default values
for both the global rule and individual entity rules.
The entity configuration will support all out of the box supported
entities, but will also allow new entities to be added directly to the
configuration file. As long as the entity name matches a name registered
in EntityList, it can be used and a rule will be created.

The default values specified in the configuration file will be used on
creation of new worlds, or when opening existing worlds for this first
time with this mod.
Users will able to set up their default mobGriefing values and have it
apply to all future worlds.

Add i18n entries for the configuration menu to allow later localization.

Close #15