Implement support for BlockFarmland mob griefing
When fallen upon by entities, BlockFarmland is reverted back to dirt if
the mobGriefing game rule is enabled.
Update BetterMobGriefingGameRuleIClassTransformer to transform the
BlockFarmland class to use BetterMobGriefingRules mobGriefing handling.
This will allow farmland block damage to be controlled on a per entity
basis.
The transformMobGreifingGameRule will need to be changed to be more
generic, the default getGameRule method is an instance and it is currently
assumed that the object initially used to retrieve that instance is an
entity (via the entities world object variable). In the case of
BlockFarmland this is not the case and the world object is used directly.
A map should be populated which maps classname to method and method to
entity index. The entity index will be the bytecode index of the entity to
get the mobGriefing value for. This will allow a more generic solution and
also allow predetermination of the entity to be used instead of trying to
calculate it at runtime.

The BlockFarmland onFallenUpon method uses an instance of the type Entity,
currently BetterMobGriefingGameRule.isMobGriefingEnabled takes an
EntityLiving parameter. This needs to be changed to Entity to allow
BlockFarmland mob griefing to be supported.

Update test classes to support the changes to method signatures and static
initialization.