Improve Wither mob griefing maintainability
The current implementation to handle Wither mob griefing uses the
EntityJoinWorldEvent and inserts a new AI package. That AI package
overrides a timer to stop block damage. However, to maintain the existing
behaviour a chunk of code was duplicated to the AI package.
When moving to future versions of Minecraft this approach requires
disproportionate amount of work due to the duplicated code having to be
checked for changes in the later iterations.
To resolve this the current handling should be removed and a Wither
target should be added to the core mod transformer. In this case it is
seen as appropriate to use the core mod due to the maintainability
concerns.
The existing handling for avoiding mob griefing caused by Wither
explosions and attacks is still usable, so only the instance of
updateAITasks mob griefing should be transformed.