Improve Sheep mob griefing maintainability
The current implementation to handle Sheep mob greifing switches out the
EntityAIEatGrass task with a custom one. However, for the custom task to
work code was duplicated from the original implementation.
When moving to future versions of Minecraft this approach requires
disproportionate amount of work to check the duplicate code for changes.

To resolve this the current handling should be removed and the necessary
changes made to the class transformer to handle modifying the original
EntityAIEatGrass AI task class.

As the EntityAIEatGrass task target entity is stored in a field a
different approach is needed for transformation. Modify the transformer to
use a list of instructions rather than variable indexes. This allows the
necesary instructions to be built in one place and then applied by the
transformer.
Update the transformMobGriefingGameRule method to insert all of the
instructions, rather than modifying just the variable index.