Refactor mob griefing command related code
Perform general refactoring of code and javadoc for mob griefing command
related code.

Delete MigratePre030BetterMobGriefingGameRule as migration will not be
supported in release 1.0.0.
Remove the calls to the migrator from BetterMobGriefingGameRule.

Rename BetterMobGriefingGameRuleWorldSavedData to EntityMobGriefingData.
Update EntityMobGriefingData, add get and set methods for mob griefing
values in the world data, the map in the world data should be made
private and only the getter and setter methods used. Create a
populateFromConfiguration method which will allow the world data to be
populated from a provided configuration, an overwrite flag should be used
to determine the behaviour for existing entities.
Rename BetterMobGriefingGameRuleWorldSavedDataTest to
EntityMobGriefingDataTest, rewrite the unit tests to use mocking instead
of faking and perform clearer assertions.

Rename BetterMobGriefingGameRuleCommandGameRule to
BetterMobGriefingComand.
Update BetterMobGriefingCommand to use the new methods added to
EntityMobGriefingData to get and set entity mob griefing values.
Rename BetterMobGriefingGameRuleCommandGameRuleTest to
BetterMobGriefingCommandTest.

Create a MobGriefingValue enumeration, which defines the valid values for
the mobGriefing game rule.