Fix configuration issue caused during refactoring
The configuration is not being initialized correctly when it does not
already exist. This is due to the configuration being read on demand and
not when the configuration is created.

Update DefaultMobGriefingConfiguration to create a synchronize method
which will read the configuration file in to fields within the
configuration class instance. The getters will retrieve the values
directly from these fields.
The synchronize method should be called from the constructor and be public
to allow it to be called when the configuration is changed.

Move the configuration message keys to
DefaultMobGriefingConfigurationConstants and delete the now empty
BetterMobGriefingGameRuleMessages class.

Remove the parent configuration category and include the global and entity
categories directly to clean up the configuration interface.

Update the relevant test classes to account for the changes made.