Add code quality gradle plugins
Update build.gradle to add FindBugs and PMD gradle plugins.

Add a FindBugs exclude filter to exclude warnings in JMockit expectations
within test classes.

Update BetterMobGriefingGameRule to convert the configuration field from a
static field to an instance field. The configuration will be initialized
on the pre-initialization event and retrieved using a new getter method.
Add a static getInstance() method to retrieve the instance of the mob from
the Loader.

Update MobGriefingEventHandler to make the explosion coordinate equality
check use Math.abs() and get the configuration using the getInstance()
method to get the entry point instance first.

Update ModInfoConstants to correct the BASE_PACKAGE string.

Update DefaultMobGriefingConfigurationConstants to make the class list an
unmodifiable list.

Update DefaultMobGriefingConfigGui to use the instance of the entry point
to get the configuration.