Update .gitignore file
Update the .gitignore file to keep up with the latest changes on the
Github sample gitignores.
Add a comment for the java compiled class file filter.
Add additional package files.
Update the gradle build folder filter to only use the local build folder,
this avoids any issues with a folder named build within the codebase.
Update the Minecraft logs and run folders to restrict to the top level
directory for the same reasons as the build folder.
Remove the code-coverage folder filter as that folder is no longer used.