Move Break Door AI task handling to core mod
It is not possible to determine whether the break door AI task has been
flagged as not to execute due to the the door interaction check or due to
global mob griefing, enforcing the entity rule may override more than
intended.

Delete BetterBreakDoorAiTask and its unit tests.

Update MobGriefingEventHandler to no longer handle the switching of AI
tasks.

Update ClassTransformer and ObfuscationHelper to allow the AI task to be
transformed by the core mod.