Handle harvest farmland mob griefing
Villagers are able to harvets farmland, this is controlled by the mob
griefing rules.

Update ClassTransformer and ObfuscationHelper to allow the
EntityAIHarvestFarmland class to be transformed by the coremod.
This has to be done by the core mod as there is otherwise no way to
override a global rule of false, as the parent should execute checks will
never run when global mob griefing is false and the runDelay triggers it
to be checked.