Implement handling of rabbit mob griefing
Update ClassTransformer and ObfuscationHelper to allow the EntityRabbit
class to be transformed by the coremod. This has to be done by the core
mod as there is otherwise no way to override a global rule of false, as
the parent should execute checks will never run when global mob griefing
is false and the runDelay trigger it to be checked.