Perform initial converstion to Minecraft 1.9.4
Bump Minecraft version to 1.9.4.
Update Forge version to latest recommended for 1.9.4.

Update build.gradle to incorporate the changes needed for version changes,
including updating the MCP mappings to stable_26.
Add the path to the FindBugs filter file for the main codebase.

Update FindBugs filters, narrow the scope of the existing test project
filer and add handling for additional false positives detected in this
version.

Update classes to account for changes in field and method naming and
visibility.