Perform initial conversion to Minecraft 1.11.2
Update README, build.gradle and ModInfoConstants to reflect the version
changes of Minecraft and Forge.
Update MCP mappings to stable_32.

Update ObfuscationHelper for the changes in field and method name
mappings.

Update ClassTransformer to update the field used by Silverfish's
AIHideInStone and to move the handling of farmland destruction from
BlockFarmland to Entity.

Update BetterMobGriefingCommand and DefaultMobgriefingConfiguration to
handle the changes made to EntityList visibility and the introduction of
internal entity names, the old names should still be used by this mod.

Update DefaultMobGriefingConfigGuiFactory to account for changes in the
parent GUI config class.