Implement handling for Evoker mob griefing
The Evoker monster's behaviour is controlled by the mobGriefing game rule
when determining whether the Evoker is able to change blue sheep to red
sheep.

Update ClassTransformer to transform the instance of mobGriefing.

Update the ConfigurationConstants to include EntityEvoker in the list of
default supported entities.

Update README.md to include the Evoker behaviour.