Fix Evoker bytecode transformation field name
When running the bytecode transformation in an obfuscated environment the
field used to determine the evoker instance is incorrect. It should use
"field_190880_e" instead of "this$0".

Update ObfuscationHelper to add a mapping between the deobfuscated and
obfuscated field names.

Update ClassTransformer to use ObfuscationHelper to determine the correct
field name.