Fix inconsistent handling of translation names
When checking whether mob griefing is enabled the translation name is
incorrectly being used, the resource path should be used to determine the
entity name to check the value of.
When setting mob griefing values via the command line translation names
which are similar to actual entity names are handled poorly. A difference
in casing will cause the entity to be identified but a new rule will be
created with the exact casing given.

Update BetterMobGriefingGameRule and BetterMobGriefingCommand to fix the
inconsistent handling.