Fix Silverfish spawning issue
When attempting to spawn Silverfish Minecraft either gives an error
message or crashes depending on the spawn method. This is caused by the
transform of AISummonSilverfish not being applied properly.

Update ClassTransformer to remove the transform of AISummonSilverfish,
it appears that the transform instructions are correct but due to
problems with the Forge transform framework the changes are not applied
correctly.
This will mean that Silverfish destroying blocks when spawning will no
longer be handled by this mod, until an alternative is found.

Update ClassTransformerTest to remove the test for AISummonSilverfish
transformation.

Update build.gradle to increment the mod version to 6.1.0 and bump the
Forge version used to 14.23.2.2611.

Update the README to note that the removed behaviour is no longer
handled and to show the new Forge version.

Close #17, close #18.