Rewrite mod to use EntityMobGriefingEvent event
The EntityMobGriefingEvent has been added to forge which allows direct
access to the griefing entity type from all instances of mob griefing.

Update MobGriefingEventHandler to remove the detonate event handling and
add handling for EntityMobGriefingEvent.
Update MobGriefingEventHandlerTest account for these changes.

Delete LoadingPlugin and ClassTransformer and the associated test
classes. A core mod is no longer needed due to the availability of the
new event.

Delete ObfuscationHelper and ObfuscationHelperTest, with no more
requirement for a core mod the number of obfuscated fields which must be
accessed will be reduced significantly to the point that this class is
superfluous.

Update EntityMobGriefingDataTest to remove unnecesarry calls to
ObfuscationHelper.

Update BetterMobGriefingGameRule and BetterMobGriefingGameRuleTest to
replace the call to ObfuscationHelper.

Update build.gradle to bump the Forge version to 14.23.2.2646 where the
new mob griefing event is available. Bump the mod version to 7.0.0.

Update ModInfoConstants to bump the mod version to 7.0.0.

Update README.md to reflect the new required Forge version and remove
the note that Silverfish are not fully handled.