Fix NullPointerException with reloaded fireballs
When a fireball becomes unloaded, such as flying in to an unloaded
chunk, a NullPointerException is thrown when the fireball is reloaded.
This occurs because the shooting entity information is lost as a result
of this process and a null value is then used to determine mob griefing
behaviour.

Update MobGriefingEventHandler to default to the global rule setting if
the entity value of the event is null.
Update MobGriefingEventHandlerTest to add a test case for this scenario.

Update build.gradle and ModInfoConstants to bump the mod version to
1.12.2-7.0.1.

Close #19