Update checkstyle config and run formatter
The checkstyle configuration is out of date, newer versions include the
Google style checks in the Jar and can be used directly instead of
copying the configuration to the repo.
Update build.gradle to use the latest checkstyle version with the google
style checks.

Delete the checkstyle configuration previously required to be provided
in the repo.